recommendation_engine/requirements.txt
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
pandas>=2.0.0
redis>=5.0.0
psycopg2-binary>=2.9.0
